eOrder: correct input argument't type and format
description: setOrderId(int orderId) --> setOrderId(long orderId)

+++ src/com/androidodc/eorder/datatypes/Order.java
@@ -28,7 +28,7 @@ public class Order implements Serializable {
         return mOrderId;
     }
 
-    public void setOrderId(int orderId) {
+    public void setOrderId(long orderId) {
         mOrderId = orderId;
     }
